:see_no_evil: Add gitignore
This is a common file for git repos, which tells git to not consider
these files/directories when it looks for changes.

This template was taken from here:
https://github.com/github/gitignore/blob/master/Python.gitignore

I also added that it should ignore svg files. You don't need to track
these since they can be generated by your code